Reorg WordPress puppet script
Move the WordPress installation stuff to the top because it's probably
going to change more and is more directly relevant. The WP-CLI stuff
and other utility ops that aren't going to change very often are moving
to the bottom.

Also, auto-installed plugins are now listed in an array. Just add the
plugin slug to the `$plugins` array to auto-install it.

Adds mrss - fixes #38