Let the VM handle SVN checkouts if the host can't
1. Ask for a WordPress.com username and password
2. Set facter facts with the SVN username and password
3. If the directories don't exist, do a checkout using the auth
that we collected (for the shared plugins repo)

This only does the SVN checkout if the directory doesn't exist.
In the case of the shared plugins repo, it also requires that
we give it a username and password.

The SVN up only happens if `svn info` returns succesful. If it's
not an SVN repo or a different version of SVN was used to do the
checkout, the exit code won't be 0 so we won't try to `svn up`.

Note: If the username and password were wrong, it's going to cause
stuf to fail. We should probably introduce better error handling

see #31, #58