Change root mysql password to really be blank instead of "blank"
It was fairly common for people to be confused by this. It's something
that was inherited from VVV. There's no reason to have a non-blank
password for the root mysql user.

Unfortunately, the module we're using won't change the password
to blank if you already have a non blank password set. There are
a couple solutions to this if you want to change it to blank.

1. `vagrant destroy` and rebuild the VM. If you have stuff that you need
in the database, back it up please :)
2. Login to mysql from the command line as root and run `SET PASSWORD for
root@localhost=PASSWORD('');`

See https://github.com/example42/puppet-mysql/issues/37