SVN: Offload WordPress checkout to the VM
Since the VM has updated SVN, we can completely offload
SVN to it and still have compatible versions between
the host SVN version and the guest.

We should leave the shared plugins in the vip-init script for now
because authentication can still cause issues if you don't
enter the correct username/password when it happens on the VM.
Ideally we can figure out how to do that on the VM in a clean way
as well.