Add HACKING.md file
The HACKING file is a way to get started developing VIP Quickstart
if you're not familiar with the project. The CONTRIBUTING file
details how you should go about getting your changes merged in.

Both of these should be considered living documents, if there's
something that's not detailed, we should be happy to add documentation
to make it as easy as possible to contribute.

Fixes #74