Adding WordPress-Tests checkout.
Added a checkout of WordPress-Tests to /vagrant/www/wp-tests and the tests configuration file to /vagrant/www/wp-tests/wp-tests-config.php.

phpunit is now installed on the vm for executing tests.

Added a `wptests` database with a matching `wptests` user to access it.

Added the environment var `WP_TESTS_DIR` to /etc/environment (using init.pp).

Closes #71