Adding dashboard widget for `RepoMonitor`.
The dashboard widget is shown on the VIP Dashboard within quickstart. The
widget shows a very basic representation of which repos are being watched,
highlighting those that are out of date.

Note that automatic repo scanning has not yet been implemented and as such
the status is from the result of the last scan initiated by wp-cli.