Add repo scan when adding repo to be monitored.
Adding a scan during the add_repo() call causes the script to first try
and check the status of the git repo before it is added. This means two
things: the scan will fail if the repo doesn't exist or is the wrong type
and that we start of with repo status info.