Adding VIPThemeHelper to Quickstart Dashboard.
The `VIPThemeHelper` is a dashboard plugin that is able to query
WordPress.com VIP for the users' VIP themes. Once it has a list of themes
it can prompt the user to install them automatically. It does this by
doing an `svn checkout` of the theme into `themes/vip/$theme`.

After a theme is checked out, it is able to both network activate and
switch the current theme to that theme for the user, making setting up
themes very simple. In most cases, the installation and activation is done
as one step.

A number of capability checks need to be added yet to keep users in-line.

See: #140, #139