Added `get_theme_action_link` helper function and cleaned up query args.
Previously there were different query args depending on whether or not a
them was supposed to be installed and/or activated. Now the theme to be
processed is passed as the `vipthemehelper-action` with option `install`
and `activate` boolean args.

The `get_theme_action_link` function generates a link that when navigated
to will cause the correct action to be taken.