Added checking for dirty working directories.
If a repo is out of date and the working directory has uncommited changes
the user is told to first commit or stash their changes before
`RepoMonitor` can update automatically.

Eventually, when automatic updating is supported, users will only be given
the update option if the working directory is clean.