Adding client side dependency validation and notices to options sync preview.
The options sync preview now gives immediate feedback to the user based on their selection of merge strategies. It checks dependencies and informs the user if a table is dependent on another which has a more restrictive merge strategy applied to it. It also automatically marks the tables that will be skipped as a result of the users' selection.