Vagrant: Shared folders: Only use vagrant shared folder on vip repo.
Vagrant shared folders are slow, but we don't really need everything
synced all the time. All the repositories that need to be kept
updated are managed from inside the VM. The only thing we really need
to keep synced with the local machine is the VIP repo because that's
where our themes and plugins live. Using Rsync for everything else
means it's nice and fast and still sync on `vagrant reload` or
if you force a sync.

See #328 #218