VIP Quickstart: Allow PHP-FPM log to be used and modified.
The default settings in `/etc/php5/fpm/pool.d/www.conf` don't allow for the log file to be changed because the use the `php_admin_value` directive.

By default the log file at `/var/log/php-fpm-www-error.log` should be working, but it doesn't exist and PHP (running as www-data) doesn't have access to create it, so we `touch` it and `chown` it.